src/core/theorem_prover.py
"""
Main theorem prover implementation.
"""
from typing import Dict, List, Any
from ..core.base import TheoremProver, ProofGenerator, ProofVerifier

class HybridTheoremProver(TheoremProver):
    """Hybrid theorem prover combining LLM and symbolic reasoning."""
    
    def __init__(self, generator: ProofGenerator, verifier: ProofVerifier):
        """Initialize the theorem prover.
        
        Args:
            generator: Proof generator instance
            verifier: Proof verifier instance
        """
        self.generator = generator
        self.verifier = verifier
        
    def prove(self, theorem: str) -> Dict[str, Any]:
        """Prove a given theorem.
        
        Args:
            theorem: The theorem to prove in string format
            
        Returns:
            Dictionary containing proof details and verification results
        """
        try:
            # Generate proof
            proof_steps = self.generator.generate_proof(theorem)
            
            # Verify final proof
            try:
                verification = self.verifier.verify_proof(theorem, proof_steps)
                
                if verification["is_valid"]:
                    return {
                        "success": True,
                        "proof": proof_steps,
                        "verification": verification
                    }
                else:
                    return {
                        "success": False,
                        "error": verification.get("error", "Failed to generate valid proof"),
                        "partial_proof": proof_steps,
                        "verification": verification
                    }
                    
            except TimeoutError as e:
                return {
                    "success": False,
                    "error": f"Verification timeout: {str(e)}",
                    "partial_proof": proof_steps,
                    "verification": None
                }
                
        except Exception as e:
            return {
                "success": False,
                "error": f"Error during proof generation: {str(e)}",
                "partial_proof": None,
                "verification": None
            }
    
    def verify_existing_proof(self, theorem: str, proof: List[str]) -> Dict[str, Any]:
        """Verify an existing proof.
        
        Args:
            theorem: The theorem being proved
            proof: List of proof steps
            
        Returns:
            Dictionary containing verification results
        """
        try:
            # Verify proof
            verification = self.verifier.verify_proof(theorem, proof)
            
            return {
                "success": verification["is_valid"],
                "verification": verification,
                "error": verification.get("error") if not verification["is_valid"] else None
            }
            
        except TimeoutError as e:
            return {
                "success": False,
                "error": f"Verification timeout: {str(e)}",
                "verification": None
            }
        except Exception as e:
            return {
                "success": False,
                "error": f"Error during proof verification: {str(e)}",
                "verification": None
            }
    
    def analyze_proof(self, theorem: str, proof: List[str]) -> Dict[str, Any]:
        """Analyze a proof for potential improvements.
        
        Args:
            theorem: The theorem being proved
            proof: List of proof steps
            
        Returns:
            Dictionary containing analysis results
        """
        try:
            # Verify current proof
            verification = self.verifier.verify_proof(theorem, proof)
            
            if not verification["is_valid"]:
                # Generate new proof if current one is invalid
                new_proof = self.generator.generate_proof(theorem)
                new_verification = self.verifier.verify_proof(theorem, new_proof)
                
                return {
                    "current_proof_valid": False,
                    "suggested_improvements": {
                        "new_proof": new_proof,
                        "verification": new_verification
                    }
                }
            
            # Try to generate a more concise proof
            new_proof = self.generator.generate_proof(theorem)
            
            if len(new_proof) < len(proof):
                return {
                    "current_proof_valid": True,
                    "suggested_improvements": {
                        "more_concise_proof": new_proof,
                        "length_reduction": len(proof) - len(new_proof)
                    }
                }
            
            return {
                "current_proof_valid": True,
                "suggested_improvements": None
            }
            
        except Exception as e:
            return {
                "error": f"Error during proof analysis: {str(e)}",
                "current_proof_valid": None,
                "suggested_improvements": None
            } 